+++ lsdviztools/lsdbasemaptools/__init__.py
@@ -8,3 +8,5 @@ from __future__ import absolute_import, division, print_function
 
 from .lsdmap_basemaptools import *
 from .lsdmap_otgrabber import *
+from .lsdmap_hbdemdownloader import *
+